requirements.txt
streamlit
openai
requests
ell-ai[sqlite]
pandas